Controller Purchase

+++ src/main/java/com/platzi/market/persistence/entity/Cliente.java
@@ -10,21 +10,16 @@ public class Cliente {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private String id;
-
     private String nombre;
-
     private String apellidos;
-
-    private Integer celular;
-
+    private Long celular;
     private String direccion;
 
     @Column (name = "correo_electronico")
     private String correoElectronico;
 
-    @OneToMany (mappedBy = "cliente")
+    @OneToMany(mappedBy = "cliente")
     private List<Compra> compras;
-
 
     public String getId() {
         return id;
@@ -50,11 +45,11 @@ public class Cliente {
         this.apellidos = apellidos;
     }
 
-    public Integer getCelular() {
+    public Long getCelular() {
         return celular;
     }
 
-    public void setCelular(Integer celular) {
+    public void setCelular(Long celular) {
         this.celular = celular;
     }
